Move indexed_read_stream to a common module.
