+++ src/storage/mod.rs
@@ -3,13 +3,7 @@ use bytes::Buf;
 use futures_core::{Future, Stream};
 use futures_lite::AsyncWrite;
 use num::{cast::FromPrimitive, CheckedSub, ToPrimitive, Unsigned};
-use std::{
-    cmp,
-    iter::Sum,
-    ops::{Deref, RangeBounds},
-};
-
-pub mod commit_log;
+use std::{iter::Sum, ops::Deref};
 
 #[async_trait(?Send)]
 pub trait AsyncIndexedRead {
@@ -21,6 +15,8 @@ pub trait AsyncIndexedRead {
 
     /// Type to index with.
     type Idx: Unsigned + CheckedSub + ToPrimitive + Ord + Copy;
+
+    async fn read(&self, idx: &Self::Idx) -> Result<Self::Value, Self::ReadError>;
 
     /// Index upper exclusive bound
     fn highest_index(&self) -> Self::Idx;
@@ -42,46 +38,6 @@ pub trait AsyncIndexedRead {
         self.has_index(idx)
             .then_some(idx)
             .and_then(|idx| idx.checked_sub(&self.lowest_index()))
-    }
-
-    async fn read(&self, idx: &Self::Idx) -> Result<Self::Value, Self::ReadError>;
-}
-
-pub async fn indexed_read_stream<'a, R, RB>(
-    indexed_read: &'a R,
-    index_bounds: RB,
-) -> impl Stream<Item = R::Value> + 'a
-where
-    RB: RangeBounds<R::Idx>,
-    R: AsyncIndexedRead,
-    R::Value: 'a,
-{
-    let (lo_min, hi_max) = (indexed_read.lowest_index(), indexed_read.highest_index());
-
-    let one = <R::Idx as num::One>::one();
-
-    let hi_max = hi_max - one;
-
-    let lo = match index_bounds.start_bound() {
-        std::ops::Bound::Included(x) => *x,
-        std::ops::Bound::Excluded(x) => *x + one,
-        std::ops::Bound::Unbounded => lo_min,
-    };
-
-    let hi = match index_bounds.end_bound() {
-        std::ops::Bound::Included(x) => *x,
-        std::ops::Bound::Excluded(x) => x.checked_sub(&one).unwrap_or(lo),
-        std::ops::Bound::Unbounded => hi_max,
-    };
-
-    let (lo, hi) = (cmp::max(lo, lo_min), cmp::min(hi, hi_max));
-
-    async_stream::stream! {
-        for index in num::range_inclusive(lo, hi) {
-            if let Ok(record) = indexed_read.read(&index).await {
-                yield record;
-            }
-        }
     }
 }
 
@@ -141,4 +97,6 @@ pub trait Storage:
         F: Future<Output = std::io::Result<T>>;
 }
 
+pub mod commit_log;
+pub mod common;
 pub mod impls;
